[docs] Update algorithm to work with v1 file
The previous algorithm supposed that first_timestamps was created from pts/time_scale, so it was always a multiple of time_scale.

But, even if it is generally the case, it isn't for v1 timestamps file.

In [this example](https://github.com/moi15moi/VideoTimestamps/blob/f9f1b64ef0c0f9e0f2b84210d325fc4080e57101/tests/test_text_file_timestamps.py#L9-L38), when we get over the frame 11, we need to calculate the time from the last frame pts.
Since the time_scale is 1000, the last frame pts is 567, but that pts is rounded. In reality, the last pts is 1700/3 = 566.666 (pts are never float, but that's just to show that it is slightly rounded).
Since we use the last frame pts to create a FPSTimestamps object, there is some rounding error since we start at the pts 567.

The solution for this is to use the last frame timestamps which is (1700/3)/1000 second as our first_timestamps to create a FPSTimestamps object.
But, (1700/3)/1000 isn't a multiple of time_scale (1000), we need to use a new formula that include the first_timestamps when it round or floor.

Example (the values are from test_text_file_timestamps.py)
Before:

$time = \text{roundingMethod}(frame \times {timescale \over fps}) \times {1 \over timescale} + first\_timestamps$

$time = \text{round}(1 \times {1000 \over 30}) \times {1 \over 1000} + { {1700 \over 3} \over 1000} = 0.59966666666667$

Now:

$time = {\text{roundingMethod}(({frame \over fps} + first\_timestamps) \times timescale) \over timescale} $

$time = {\text{round}(({1 \over 30} + { {1700 \over 3} \over 1000}) \times 1000) \over 1000} = 0.6$